Refactored Google Helpers to be inside app_helpers as it's functionality is now used accross pages and config
