+++ src/aims_ui/page_helpers/security_utils.py
@@ -1,7 +1,7 @@
 import re
 from aims_ui.page_controllers.f_error_pages.page_error import page_error
 from aims_ui.page_helpers.error.error_utils import error_page_no_access
-from .google_utils import get_current_group
+from ..app_helpers.google_utils import get_current_group
 
 
 def check_user_has_access_to_page(page_name):
