+++ version.txt
@@ -1,1 +1,1 @@
-1.3.76
+1.3.77
